--- code/cell.modules/pkg.sys/ui.dev/src/components/ActionPanel/View.Group.tsx
@@ -1,33 +1,0 @@
-import React from 'react';
-
-import { css, CssValue, t } from '../../common';
-import { Item } from './View.Item';
-import { Title } from './View.Title';
-
-export type GroupProps = {
-  items: t.ActionItem[];
-  name?: string;
-  style?: CssValue;
-  onItemClick?: t.ActionItemClickEventHandler;
-};
-
-export const Group: React.FC<GroupProps> = (props) => {
-  const styles = {
-    base: css({}),
-    body: css({ fontSize: 14 }),
-    list: css({}),
-  };
-
-  const elItems = props.items.map((item, i) => {
-    return <Item key={i} model={item} onClick={props.onItemClick} />;
-  });
-
-  return (
-    <div {...css(styles.base, props.style)}>
-      <div {...styles.body}>
-        {props.name && <Title>{props.name}</Title>}
-        <div {...styles.list}>{elItems}</div>
-      </div>
-    </div>
-  );
-};
