+++ code/cell.modules/pkg.sys/ui.dev/src/components/ActionPanel/View.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { constants, css, defaultValue, t, time } from '../../common';
-import { Group } from './View.Group';
+import { Item } from './Item';
 
 export type ViewProps = t.ActionPanelProps & {
   model: t.ActionModel<any>;
@@ -16,25 +16,33 @@ export const View: React.FC<ViewProps> = (props) => {
     base: css({
       Scroll: scrollable,
       overflowY: scrollable ? 'scroll' : 'hidden',
-      fontFamily: constants.FONT.SANS,
-      color: constants.COLORS.DARK,
       userSelect: 'none',
       boxSizing: 'border-box',
       paddingBottom: 50,
+      fontFamily: constants.FONT.SANS,
+      fontSize: 14,
+      color: constants.COLORS.DARK,
     }),
   };
 
   const onItemClick: t.ActionItemClickEventHandler = (e) => {
     const ctx = props.getContext ? props.getContext() : undefined;
     time.delay(0, () => {
-      // NB: Delay for a tick to make focus related actions simpler within the handler.
+      // NB:  Delay for a tick to make focus related actions simpler to
+      //      deal with within the handler.
+      //      If we do this here, then it does not have to be propogated across
+      //      any handler that cares about this.
       if (e.model.type === 'button' && e.model.onClick) e.model.onClick(ctx);
     });
   };
 
+  const elItems = model.items.map((item, i) => {
+    return <Item key={i} model={item} onClick={onItemClick} />;
+  });
+
   return (
-    <div {...css(styles.base, props.style)} className={'uih-ActionPanel'}>
-      <Group items={model.items} onItemClick={onItemClick} />
+    <div {...css(styles.base, props.style)} className={'dev-ActionPanel'}>
+      {elItems}
     </div>
   );
 };
