+++ code/cell.modules/pkg.sys/ui.dev/src/types/index.ts
@@ -1,5 +1,7 @@
 export * from './types.Host';
+
 export * from './types.ActionPanel';
 export * from './types.Actions.model';
+export * from './types.Actions.model.items';
 export * from './types.Actions.builder';
 export * from './types.Actions.factory';
